Code cleanup part 2
remove unnessacry explicit types (<"xxx">), unneccessary (un)boxing
remove redunant code (string.tostring(), stringbuffer, type casts)

+++ app/src/main/java/me/ccrama/redditslide/Views/TouchImageView.java
@@ -1123,7 +1123,7 @@ public class TouchImageView extends ImageView {
                 minY = maxY = startY;
             }
 
-            scroller.fling(startX, startY, (int) velocityX, (int) velocityY, minX,
+            scroller.fling(startX, startY, velocityX, velocityY, minX,
                     maxX, minY, maxY);
             currX = startX;
             currY = startY;
